Replaced absolute_path function with realpath
realpath is a unix tool that does the job of the ghetto absolute_path
function, but does it properly everytime instead.

THe issue with absolute_path was that it was failing to decipher
relative paths that ended with `..`, as it misinterpret that as the
basename instead of part of the directory name.

For that generation of the TARGET_PATHS_GREP, modified each given item
to go through realpath instead of being it given raw.