Updated finding of files with non absolute paths
Many files listed in dependency files may have just relative paths based
on where the compile command was called.

If the file cannot be found immediately, the script now recurses back
until it reaches the root directory, trying to find where the file is.
If it can't be found, then the script errors out instead.